Levels 0, 1, and 2 of the python challenge
In level 2 where I was supposed to find the least common characters, I
had some trouble with python reading the .txt file input correctly. I
ended up asking for user input and then pasting in the text to the
program that way. In the future I will use the method of getting the
program to read the file off of the source page directly.